Dependency Confusion Attack Deu to Unclamed Package
Summary:
clearpool.finance package, with the name "defillama-apy-serverl" found on https://github.com/clearpool-finance/defiLlama-yield-server/blob/master/package.json, is unclaimed on the NPM package. This leaves the package vulnerable to a Dependency Confusion Attack. An attacker can claim the package and upload malicious code under that unclaimed package. This results in achieving remote code execution on developers/users' machines who depend on this package.

Vulnerability Description:
The package "defillama-apy-server " found on https://github.com/clearpool-finance/defiLlama-yield-server/blob/master/package.json has not been claimed on the NPM website. This means that anyone can claim this package and upload a malicious code under that package name. When developers or users depend on this package in their code, they may unknowingly download the malicious code, leading to remote code execution on their machines.

Impact:
This vulnerability allows an attacker to execute arbitrary code on a user's machine. This can lead to a complete takeover of the user's system, stealing sensitive information, or deploying ransomware on the system.

Steps to Reproduce:

Go to https://www.npmjs.com/ and search for the "defillama-apy-server " package.
Observe that the package is unclaimed.
Claim the package under your account.
Upload a malicious code to the package.
Wait for developers or users to depend on this package in their code.
Observe the remote code execution on the users' machines.

Recommended Mitigation:
clearpool.finance should immediately claim the package "defillama-apy-server " on the NPM website to prevent any unauthorized access. They should also perform a security audit of their code to ensure that it is not vulnerable to any other dependency confusion attacks.

Developers and users who depend on the "defillama-apy-server " package should be advised to verify the authenticity of the package before adding it as a dependency. They should also avoid using unclaimed packages in their code.

POC:-
Here I have found the unclaimed package name
image.png

image.pngv

Reference:- https://dhiyaneshgeek.github.io/web/security/2021/09/04/dependency-confusion/
https://github.com/code-423n4/2022-03-timeswap-findings/issues/9 

Thanks
Darshan